[tlinh2110@gmail.com] BAEL 1313 - Update unit test
