+++ spring-security-mvc-boot/src/main/java/org/baeldung/acl/persistence/dao/NoticeMessageRepository.java
@@ -19,6 +19,6 @@ public interface NoticeMessageRepository extends JpaRepository<NoticeMessage, Lo
     
     @SuppressWarnings("unchecked")
     @PreAuthorize("hasPermission(#noticeMessage, 'WRITE')")
-    NoticeMessage save(@Param("noticeMessage")NoticeMessage chatMessage);
+    NoticeMessage save(@Param("noticeMessage")NoticeMessage noticeMessage);
 
 }
