+++ spring-security-mvc-boot/src/main/resources/acl-data.sql
@@ -8,13 +8,12 @@ INSERT INTO acl_class (id, class) VALUES
 INSERT INTO acl_sid (id, principal, sid) VALUES
 (1, 1, 'manager'),
 (2, 1, 'hr'),
-(3, 1, 'employee'),
-(4, 0, 'ROLE_EDITOR'),
-(5, 0, 'ROLE_VIEWER');
+(3, 1, 'admin'),
+(4, 0, 'ROLE_EDITOR');
 
 INSERT INTO acl_object_identity (id, object_id_class, object_id_identity, parent_object, owner_sid, entries_inheriting) VALUES
-(1, 1, 1, NULL, 1, 0),
-(2, 1, 2, NULL, 2, 0),
+(1, 1, 1, NULL, 3, 0),
+(2, 1, 2, NULL, 3, 0),
 (3, 1, 3, NULL, 3, 0)
 ;
 
@@ -25,6 +24,5 @@ INSERT INTO acl_entry (id, acl_object_identity, ace_order, sid, mask, granting, 
 (4, 2, 1, 2, 1, 1, 1, 1),
 (5, 2, 2, 4, 1, 1, 1, 1),
 (6, 3, 1, 4, 1, 1, 1, 1),
-(7, 3, 2, 4, 2, 1, 1, 1),
-(8, 3, 3, 5, 1, 1, 1, 1)
+(7, 3, 2, 4, 2, 1, 1, 1)
 ;
